+++ templates/ts-react-app/src/services/restApi/queries/baseApi.service.ts
@@ -1,6 +1,6 @@
 import { createApi } from '@reduxjs/toolkit/dist/query/react/index';
-import { baseQueryWithTokenExpirationCheck } from './config/query.config';
-import { TAG_TYPES } from './config/apiTags';
+import { baseQueryWithTokenExpirationCheck } from '../config/query.config';
+import { TAG_TYPES } from '../config/apiTags';
 
 export const baseApi = createApi({
   reducerPath: 'baseApi',